// client/src/interfaces/strings.ts
export interface IStringBundle {
    // navigation ========================================
    navItemHome: string;
    navItemDiscover: string;
    navItemMyQuizzes: string;
    navItemLeaderboard: string;
    navItemSearchPlaceholder: string;
    navItemSkipToMainContent: string;

    navItemProfileMenuSR: string;
    navItemProfileMenuMyProfile: string;
    navItemProfileMenuSettings: string;
    navItemProfileMenuSignout: string;
    navItemProfileMenuSignin: string;

    navItemBurgerMenySR: string;
    navItemBurgerMenu: string;
    navItemBurgerOptionCreateQuiz: string;
    navItemBurgerOptionCreateQuizDescription: string;
    navItemBurgerOptionPlaylists: string;
    navItemBurgerOptionPlaylistsDescription: string;

    navItemToggleTheme: string;
    navItemSelectLanguage: string;

    // Footer ===========================================
    footerTermsOfService: string;
    footerPrivacyPolicy: string;
    footerPackagesAndLicenses: string;
    footerCredits: string;
    footerCopyRight: string;

    // promote banner ===================================
    promoteBannerTitle: string;
    promoteBannerDescription: string;
    promoteBannerButtonText: string;
    promoteBannerDismissSR: string;

    // Quiz Slideover ===================================
    quizSlideoverTitle: string;
    quizSlideoverDescription: string;
    quizSlideoverClosePanelSR: string;
    quizSlideoverTitleLabel: string;
    quizSlideoverTitlePlaceholder: string;
    quizSlideoverDescriptionLabel: string;
    quizSlideoverDescriptionPlaceholder: string;
    quizSlideoverManageContributorsTitle: string;
    quizSlideoverAddContributorsSR: string;
    quizSlideoverPrivateFieldLabel: string;
    quizSlideoverPrivateFieldDescription: string;
    quizSlideoverPublicOptionLabel: string;
    quizSlideoverPublicOptionDescription: string;
    quizSlideoverPrivateOptionLabel: string;
    quizSlideoverPrivateOptionDescription: string;
    quizSlideoverCopyLinkButton: string;
    quizSlideoverCancelButton: string;
    quizSlideoverNextButton: string;

    quizSlideoverErrorTitle: string;
    quizSlideoverErrorDescription: string;

    // home ==============================================
    homePageBannerTitle: string;
    homePageBannerLink: string;
    homePageTitle: string;
    homePageDescription: string;
    homePageGameInputSR: string;
    homePageGameInputPlaceholder: string;
    homePageGameInputButton: string;
    homePageGameInputError: string;
    homePageDiscoverMore: string;

    // Leaderboard =========================================
    leaderboardTitle: string;
    leaderboardSubTitle: string;
    leaderboardFilterSR: string;
    leaderboardFilterOptionMyConnections: string;
    leaderboardFilterOptionAllTime: string;
    leaderboardFilterOptionMonthly: string;
}